+++ apps/app/src/components/login/index.ts
@@ -1,2 +1,3 @@
 export * from './KakaoLogin'
+export * from './GoogleLogin'
 export * from './OAuthButton'
